requirements.txt
# Python dependencies for Pub/Sub to GCS microservice :)
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0
google-auth>=2.17.0

# Additional utilities for better service operation :)
python-dotenv>=1.0.0  # For local development with .env files :)
structlog>=23.1.0     # For structured logging :)